feat(worker): wire Claude→Gemini translation fallback
Add a LyricsWorker::build_gemini_clients helper that produces the
per-song Vec<GeminiClient> (split gemini_api_key on commas, drop
empties, wrap each key with max_attempts=1). Replace the inline
block in process_song with a call to this helper so alignment and
translation share the same client list.

Add LyricsWorker::translate_with_fallback and apply_translations:
translate_track now tries Claude first and falls back to Gemini on
ANY Claude error (including the "copyrighted material" policy
refusal). retry_missing_translations uses the same fallback so a
song that ships without Slovak due to a Claude refusal is retried
against Gemini on the next retry tick.

To keep worker.rs under the 1000-line cap, move gather_sources_impl
into a sibling gather module. worker.rs re-exports it as
crate::lyrics::worker::gather_sources_impl so worker_tests.rs
imports keep working unchanged.
