+++ crates/sp-server/src/lyrics/worker.rs
@@ -4,7 +4,8 @@
 //!   1. gather_sources: YT manual subs + LRCLIB + autosub json3 in parallel.
 //!   2. Vocal isolation (Qwen3 provider; best-effort).
 //!   3. Orchestrator::process_song → LyricsTrack.
-//!   4. SK translation (Claude only; Gemini quota reserved for alignment).
+//!   4. SK translation — Claude (CLIProxyAPI) first, Gemini direct-API
+//!      fallback when Claude refuses on copyright policy.
 //!   5. Persist JSON + DB row with pipeline_version + quality_score.
 
 use anyhow::Result;
@@ -21,7 +22,7 @@ use tracing::{debug, error, info, warn};
 use crate::{
     ai::client::AiClient,
     db::models::get_next_video_missing_translation,
-    lyrics::{aligner, genius, lrclib, translator, youtube_subs},
+    lyrics::{aligner, translator},
 };
 
 #[allow(dead_code)]
@@ -55,203 +56,10 @@ struct RetryBackoff {
     consecutive_failures: u32,
 }
 
-/// Free function containing the `gather_sources` logic so it can be tested
-/// without constructing a full `LyricsWorker`.
-///
-/// mutants::skip: legacy LRCLIB guards (lines 99-100) + description match guard are
-/// exercised end-to-end by `gather_sources_pushes_description_candidate_when_claude_returns_lyrics`
-/// and `gather_sources_skips_description_when_claude_returns_empty_array` integration tests
-/// (plus the structural call-order test further down); individual mutations in these
-/// I/O-bound branches cannot be killed by unit tests without a full mock harness for
-/// yt-dlp/LRCLIB/autosub, which is out of scope.
-#[cfg_attr(test, mutants::skip)]
-#[allow(clippy::too_many_arguments)]
-pub(crate) async fn gather_sources_impl(
-    ai_client: Option<&crate::ai::client::AiClient>,
-    ytdlp_path: &std::path::Path,
-    cache_dir: &std::path::Path,
-    client: &reqwest::Client,
-    row: &crate::db::models::VideoLyricsRow,
-    autosub_tmp_dir: &std::path::Path,
-    genius_access_token: &str,
-) -> Result<crate::lyrics::provider::SongContext> {
-    use crate::lyrics::autosub_provider::fetch_autosub;
-    use crate::lyrics::provider::{CandidateText, SongContext};
-
-    let youtube_id = row.youtube_id.clone();
-    let audio_path = row
-        .audio_file_path
-        .as_ref()
-        .map(PathBuf::from)
-        .unwrap_or_default();
-
-    // 1. Manual yt_subs
-    let yt_tmp = std::env::temp_dir().join("sp_yt_subs");
-    let _ = tokio::fs::create_dir_all(&yt_tmp).await;
-    let yt_subs_track = match youtube_subs::fetch_subtitles(ytdlp_path, &youtube_id, &yt_tmp).await
-    {
-        Ok(Some(track)) => {
-            info!("gather: YT manual subs hit for {youtube_id}");
-            Some(track)
-        }
-        Ok(None) => {
-            debug!("gather: no YT manual subs for {youtube_id}");
-            None
-        }
-        Err(e) => {
-            warn!("gather: YT sub fetch error for {youtube_id}: {e}");
-            None
-        }
-    };
-
-    // 2. LRCLIB (if song/artist known)
-    let lrclib_track = if !row.song.is_empty() && !row.artist.is_empty() {
-        let duration_s = row.duration_ms.map(|ms| (ms / 1000) as u32).unwrap_or(0);
-        match lrclib::fetch_lyrics(client, &row.artist, &row.song, duration_s).await {
-            Ok(Some(track)) => {
-                info!("gather: LRCLIB hit for {youtube_id}");
-                Some(track)
-            }
-            Ok(None) => None,
-            Err(e) => {
-                warn!("gather: LRCLIB error for {youtube_id}: {e}");
-                None
-            }
-        }
-    } else {
-        None
-    };
-
-    // 2a. Genius (documented API + public lyric-page scrape). No timing.
-    let genius_track = if !row.song.is_empty() && !row.artist.is_empty() {
-        match genius::fetch_lyrics(client, genius_access_token, &row.artist, &row.song).await {
-            Ok(Some(t)) => {
-                info!(%youtube_id, line_count = t.lines.len(), "gather: Genius hit");
-                Some(t)
-            }
-            Ok(None) => None,
-            Err(e) => {
-                warn!("gather: Genius error for {youtube_id}: {e}");
-                None
-            }
-        }
-    } else {
-        None
-    };
-
-    // 3. Auto-sub json3
-    let autosub_json3 = match fetch_autosub(ytdlp_path, &youtube_id, autosub_tmp_dir).await {
-        Ok(Some(p)) => Some(p),
-        Ok(None) => None,
-        Err(e) => {
-            warn!("gather: autosub fetch error for {youtube_id}: {e}");
-            None
-        }
-    };
-
-    let mut candidate_texts: Vec<CandidateText> = Vec::new();
-
-    // 0. Operator-provided override (V15). Highest priority — when an
-    //    operator has pasted lyrics for a song, they expect those lines
-    //    to drive alignment, not whatever yt_subs/LRCLIB/description the
-    //    gather paths produce. Empty/whitespace override is ignored.
-    if let Some(raw) = row.lyrics_override_text.as_ref() {
-        let lines: Vec<String> = raw
-            .lines()
-            .map(|s| s.trim().to_string())
-            .filter(|s| !s.is_empty())
-            .collect();
-        if !lines.is_empty() {
-            info!(
-                youtube_id = %youtube_id,
-                line_count = lines.len(),
-                "gather: operator lyrics override present"
-            );
-            candidate_texts.push(CandidateText {
-                source: "override".into(),
-                lines,
-                has_timing: false,
-                line_timings: None,
-            });
-        }
-    }
-    if let Some(t) = &yt_subs_track {
-        candidate_texts.push(CandidateText {
-            source: "yt_subs".into(),
-            lines: t.lines.iter().map(|l| l.en.clone()).collect(),
-            has_timing: true,
-            line_timings: Some(t.lines.iter().map(|l| (l.start_ms, l.end_ms)).collect()),
-        });
-    }
-    if let Some(t) = &lrclib_track {
-        candidate_texts.push(CandidateText {
-            source: "lrclib".into(),
-            lines: t.lines.iter().map(|l| l.en.clone()).collect(),
-            has_timing: true,
-            line_timings: Some(t.lines.iter().map(|l| (l.start_ms, l.end_ms)).collect()),
-        });
-    }
-    if let Some(t) = &genius_track {
-        candidate_texts.push(CandidateText {
-            source: "genius".into(),
-            lines: t.lines.iter().map(|l| l.en.clone()).collect(),
-            has_timing: false,
-            line_timings: None,
-        });
-    }
-
-    // 4. YouTube description lyrics (LLM-extracted). Best-effort.
-    if let Some(ai) = ai_client {
-        let description_lines = match crate::lyrics::description_provider::fetch_description_lyrics(
-            ai,
-            ytdlp_path,
-            &youtube_id,
-            cache_dir,
-            &row.song,
-            &row.artist,
-        )
-        .await
-        {
-            Ok(Some(lines)) if !lines.is_empty() => {
-                info!(
-                    youtube_id = %youtube_id,
-                    line_count = lines.len(),
-                    "gather: description lyrics hit"
-                );
-                Some(lines)
-            }
-            Ok(_) => {
-                debug!("gather: no description lyrics for {youtube_id}");
-                None
-            }
-            Err(e) => {
-                warn!("gather: description fetch error for {youtube_id}: {e}");
-                None
-            }
-        };
-        if let Some(lines) = description_lines {
-            candidate_texts.push(CandidateText {
-                source: "description".into(),
-                lines,
-                has_timing: false,
-                line_timings: None,
-            });
-        }
-    }
-
-    if candidate_texts.is_empty() {
-        anyhow::bail!("no text sources available for {youtube_id}");
-    }
-
-    Ok(SongContext {
-        video_id: youtube_id,
-        audio_path,
-        clean_vocal_path: None, // filled by process_song before orchestrator call
-        candidate_texts,
-        autosub_json3,
-        duration_ms: row.duration_ms.unwrap_or(0) as u64,
-    })
-}
+/// Re-export so `worker_tests` can keep importing from
+/// `crate::lyrics::worker::gather_sources_impl`. The body lives in the
+/// sibling `gather` module so `worker.rs` stays under the 1000-line cap.
+pub(crate) use crate::lyrics::gather::gather_sources_impl;
 
 impl LyricsWorker {
     #[allow(clippy::too_many_arguments)]
@@ -493,28 +301,75 @@ impl LyricsWorker {
         .await
     }
 
-    /// EN→SK translation via Claude. No Gemini fallback — Gemini quota is
-    /// reserved for alignment (which dominates call volume). If Claude fails,
-    /// the track ships without translation; the UI degrades gracefully.
+    /// Build per-song Vec<GeminiClient>: split `gemini_api_key` by comma, drop
+    /// empties, wrap each key in a direct-API client with `max_attempts=1`.
+    // mutants::skip: pure split/trim/filter over config; covered by gemini_provider tests.
+    #[cfg_attr(test, mutants::skip)]
+    fn build_gemini_clients(&self) -> Vec<crate::lyrics::gemini_client::GeminiClient> {
+        use crate::lyrics::gemini_client::GeminiClient;
+        let model = std::env::var("GEMINI_LYRICS_MODEL")
+            .unwrap_or_else(|_| "gemini-3.1-pro-preview".to_string());
+        self.gemini_api_key
+            .split(',')
+            .map(|s| s.trim().to_string())
+            .filter(|s| !s.is_empty())
+            .map(|k| {
+                let mut c = GeminiClient::direct(k, model.clone());
+                c.max_attempts = 1;
+                c
+            })
+            .collect()
+    }
+
+    /// Claude first, Gemini fallback on ANY Claude error. Returns per-line
+    /// translations on success or an error naming both failures.
+    #[cfg_attr(test, mutants::skip)]
+    async fn translate_with_fallback(
+        &self,
+        ai_client: &AiClient,
+        track: &LyricsTrack,
+        youtube_id: &str,
+    ) -> Result<Vec<String>> {
+        let claude_err = match translator::translate_via_claude(ai_client, track).await {
+            Ok(t) => return Ok(t),
+            Err(e) => e,
+        };
+        warn!(
+            "worker: Claude translation failed for {youtube_id}: {claude_err} — falling back to Gemini"
+        );
+        let clients = self.build_gemini_clients();
+        translator::translate_via_gemini(&clients, track)
+            .await
+            .map_err(|gemini_err| {
+                anyhow::anyhow!("Claude + Gemini both failed for {youtube_id} (claude: {claude_err}, gemini: {gemini_err})")
+            })
+    }
+
+    /// Apply per-line translations to `track`. Empty strings leave `sk = None`.
+    #[cfg_attr(test, mutants::skip)]
+    fn apply_translations(track: &mut LyricsTrack, translations: Vec<String>) {
+        for (line, sk_text) in track.lines.iter_mut().zip(translations) {
+            line.sk = if sk_text.is_empty() {
+                None
+            } else {
+                Some(sk_text)
+            };
+        }
+        track.language_translation = "sk".into();
+    }
+
+    /// EN→SK step of `process_song`. Silent on failure — UI degrades.
     #[cfg_attr(test, mutants::skip)]
     async fn translate_track(&self, track: &mut LyricsTrack, youtube_id: &str) {
         let Some(ai_client) = &self.ai_client else {
             return;
         };
-        match translator::translate_via_claude(ai_client, track).await {
-            Ok(translations) => {
-                for (line, sk_text) in track.lines.iter_mut().zip(translations) {
-                    line.sk = if sk_text.is_empty() {
-                        None
-                    } else {
-                        Some(sk_text)
-                    };
-                }
-                track.language_translation = "sk".into();
-            }
-            Err(e) => {
-                warn!("worker: Claude translation failed for {youtube_id}: {e}");
-            }
+        match self
+            .translate_with_fallback(ai_client, track, youtube_id)
+            .await
+        {
+            Ok(translations) => Self::apply_translations(track, translations),
+            Err(e) => warn!("worker: translation failed for {youtube_id}: {e}"),
         }
     }
 
@@ -678,27 +533,12 @@ impl LyricsWorker {
                 "ffmpeg"
             };
             let ffmpeg_path = self.tools_dir.join(ffmpeg_name);
-            let model = std::env::var("GEMINI_LYRICS_MODEL")
-                .unwrap_or_else(|_| "gemini-3.1-pro-preview".to_string());
-            // v14: direct API with one client per key. The `gemini_api_key`
-            // DB setting is a comma-separated list of API keys; empty entries
-            // and whitespace are stripped. `max_attempts=1` per client so a
-            // 429 fails fast and the provider-level rotation in
+            // v14: direct API with one client per key. `build_gemini_clients`
+            // splits `self.gemini_api_key` on commas, strips whitespace,
+            // drops empty entries, and sets `max_attempts=1` per client so
+            // a 429 fails fast and the provider-level rotation in
             // `transcribe_rotating` advances to the next key immediately.
-            let keys: Vec<String> = self
-                .gemini_api_key
-                .split(',')
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty())
-                .collect();
-            let clients: Vec<GeminiClient> = keys
-                .into_iter()
-                .map(|k| {
-                    let mut c = GeminiClient::direct(k, model.clone());
-                    c.max_attempts = 1;
-                    c
-                })
-                .collect();
+            let clients: Vec<GeminiClient> = self.build_gemini_clients();
             if !clients.is_empty() {
                 tracing::info!(
                     "gemini: registered {} API key(s) for alignment rotation",
@@ -916,20 +756,15 @@ impl LyricsWorker {
         };
         info!("lyrics_worker: retrying translation for {youtube_id}");
 
-        // Claude-only. Gemini quota is reserved for alignment.
         let Some(ai_client) = &self.ai_client else {
             return;
         };
-        let result = match translator::translate_via_claude(ai_client, &track).await {
-            Ok(translations) => {
-                for (line, sk_text) in track.lines.iter_mut().zip(translations) {
-                    line.sk = if sk_text.is_empty() {
-                        None
-                    } else {
-                        Some(sk_text)
-                    };
-                }
-                track.language_translation = "sk".to_string();
+        let translations = self
+            .translate_with_fallback(ai_client, &track, &youtube_id)
+            .await;
+        let result: Result<()> = match translations {
+            Ok(t) => {
+                Self::apply_translations(&mut track, t);
                 Ok(())
             }
             Err(e) => Err(e),
